refactor arguments folder + run ruff
